requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pandas==0.24.2
Pillow==7.0.0
pyTelegramBotAPI==3.6.7
requests==2.23.0
selenium==3.141.0
six==1.14.0
tabulate==0.8.2
urllib3==1.25.9
WeasyPrint==51
webdriver-manager==2.3.0
Werkzeug==1.0.0
